Add authors fix to individual lip pages
