+++ src/DocComponents/LIPTable/index.js
@@ -4,6 +4,7 @@ import PropTypes from 'prop-types';
 import useDocusaurusContext from '@docusaurus/useDocusaurusContext';
 import useBaseUrl from '@docusaurus/useBaseUrl';
 
+import Author from 'src/DocComponents/Author';
 import filesByType from '../../../LIPMetadata.json';
 import {LIP_TYPE, LIP_STATUS} from '../../enums';
 import {parseAuthors} from 'src/utils';
@@ -14,23 +15,6 @@ const getLips = (status, types) =>
   types.reduce((lips, type) => {
     return lips.concat(filesByType[type][status]);
   }, []).sort((a, b) => a.lip < b.lip ? 1 : -1);
-
-const Author = ({index, title, username}) => {
-  const formattedTitle = `${index > 0 ? ', ' : ''}${title}`;
-
-  if (!username) {
-    return formattedTitle;
-  }
-
-  return (
-    <a
-      href={`https://github.com/${username}`}
-      target="_blank"
-    >
-      {formattedTitle}
-    </a>
-  );
-};
 
 const LIPRow = ({ authors, num, title }) => {
   const {siteConfig: {themeConfig}} = useDocusaurusContext();
